Display top suggestion from useSuggestions, not AICamera
