+++ src/components/Match/calculateConfidence.js
@@ -1,0 +1,8 @@
+const calculateConfidence = suggestion => {
+  if ( !suggestion ) { return null; }
+  return suggestion?.score
+    ? Math.round( suggestion.score )
+    : Math.round( suggestion.combined_score );
+};
+
+export default calculateConfidence;
